requirements.txt
streamlit==1.54.0
python-dotenv>=1.0.1
Authlib
certifi>=2024.0.0
